requirements.txt
adal==1.2.6
applicationinsights==0.11.9
argcomplete==1.12.2
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1613037097816/work
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
azure-cli-core==2.20.0
azure-cli-telemetry==1.0.6
azure-common==1.1.26
azure-core==1.12.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-core==1.2.2
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==12.1.0
azure-mgmt-storage==11.2.0
azureml-automl-core==1.24.0
azureml-contrib-dataset==1.24.0
azureml-core==1.24.0.post2
azureml-dataprep==2.11.2
azureml-dataprep-native==30.0.0
azureml-dataprep-rslex==1.9.1
azureml-dataset-runtime==1.24.0
azureml-pipeline==1.24.0
azureml-pipeline-core==1.24.0
azureml-pipeline-steps==1.24.0
azureml-sdk==1.24.0
azureml-telemetry==1.24.0
azureml-train==1.24.0
azureml-train-automl-client==1.24.0
azureml-train-core==1.24.0
azureml-train-restclients-hyperdrive==1.24.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246945912/work
chardet==4.0.0
cloudpickle==1.6.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.3.2
cycler==0.10.0
Cython @ file:///tmp/build/80754af9/cython_1614014838717/work
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
distro==1.5.0
docker==4.4.4
dotnetcore2==2.1.20
entrypoints==0.3
fusepy==3.0.1
humanfriendly==9.1
idna==2.10
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596207638929/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1614615245721/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1610481889018/work
isodate==0.6.0
jedi==0.17.0
jeepney==0.6.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
jmespath==0.10.0
jsonpickle==2.0.0
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1612213311222/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282420641/work
knack==0.8.0rc2
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1613407855456/work
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
msal==1.10.0
msal-extensions==0.2.2
msrest==0.6.21
msrestazure==0.6.4
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914830498/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
ndg-httpsclient==0.5.1
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///tmp/build/80754af9/notebook_1611340975709/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.2.3
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120460739/work
paramiko==2.7.2
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pathspec==0.8.1
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1615224089586/work
pkginfo==1.7.0
portalocker==1.7.1
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
psutil==5.8.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyarrow==1.0.1
pyasn1==0.4.8
pycocotools==2.0.2
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141720057/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz==2021.1
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole @ file:///tmp/build/80754af9/qtconsole_1612458529756/work
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
SecretStorage==3.3.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1605205327372/work
tabulate==0.8.9
terminado==0.9.2
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
torch==1.8.0
torchaudio==0.8.0a0+a751e1d
torchvision==0.9.0
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm==4.59.0
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3==1.26.3
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
websocket-client==0.58.0
widgetsnbextension==3.5.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
